Split CachedResolver to own file per jmartin-r7
